+++ CHANGELOG.md
@@ -1,10 +1,32 @@
-# :bug: Bug Fixes
+# Important Changes
 
-- Fix file search by package name [`7621fec`](https://github.com/keindev/figma-portal/commit/7621fec3947a6b0539b0eefbadbc27c778d0a78e)
+## Dependencies
 
-# :memo: Internal changes
+<details>
+<summary>Dependencies</summary>
 
-- Fix install section [`c5d8745`](https://github.com/keindev/figma-portal/commit/c5d8745d96033d1cdf36bfcd853a9e7af3555796)
+- Added **[exec-buffer](https://www.npmjs.com/package/exec-buffer/v/3.2.0)** with `^3.2.0`
+- Added **[jpegtran-bin](https://www.npmjs.com/package/jpegtran-bin/v/6.0.1)** with `^6.0.1`
+- Added **[optipng-bin](https://www.npmjs.com/package/optipng-bin/v/8.1.0)** with `^8.1.0`
+- Added **[svgo](https://www.npmjs.com/package/svgo/v/2.8.0)** with `^2.8.0`
+- Removed **[imagemin-jpegtran](https://www.npmjs.com/package/imagemin-jpegtran/v/7.0.0)**, with `^7.0.0`
+- Removed **[imagemin-optipng](https://www.npmjs.com/package/imagemin-optipng/v/8.0.0)**, with `^8.0.0`
+- Removed **[imagemin-svgo](https://www.npmjs.com/package/imagemin-svgo/v/10.0.1)**, with `^10.0.1`
+
+</details>
+
+<details>
+<summary>Dev Dependencies</summary>
+
+- Added **[@types/svgo](https://www.npmjs.com/package/@types/svgo/v/2.6.0)** with `^2.6.0`
+- Added **[ts-node](https://www.npmjs.com/package/ts-node/v/10.4.0)** with `^10.4.0`
+- Bumped **[@tagproject/ts-package-shared-config](https://www.npmjs.com/package/@tagproject/ts-package-shared-config/v/6.3.1)** from `6.2.1` to `6.3.1`
+- Bumped **[cspell](https://www.npmjs.com/package/cspell/v/5.14.0)** from `5.13.4` to `5.14.0`
+- Removed **[@types/imagemin-jpegtran](https://www.npmjs.com/package/@types/imagemin-jpegtran/v/5.0.1)**, with `^5.0.1`
+- Removed **[@types/imagemin-optipng](https://www.npmjs.com/package/@types/imagemin-optipng/v/5.2.1)**, with `^5.2.1`
+- Removed **[@types/imagemin-svgo](https://www.npmjs.com/package/@types/imagemin-svgo/v/10.0.0)**, with `^10.0.0`
+
+</details>
 
 ---